refactor(plugins): Move all `ALL` properties to `Plugin` implementations
For some plugin types, the conventional `ALL` property in the was not
defined in the actual plugin implementation, but in a class managing the
plugin implementations. Resolve this inconsistency by generall defining
the `ALL` properties in the belonging plugin implementation to prepare
for an upcoming change that inspects plugin implementations.

Signed-off-by: Sebastian Schuberth <sebastian@doubleopen.org>